refactor(ui/data): extract data regions and harden cell-action confirm
Address CI file-size ratchet and PR review feedback on #2155:

- Move CoverageGapsRegion and DataQualityRegion (plus the capability-matrix
  reveal helper) out of the oversized data-screen.tsx into a focused
  data-screen.data-regions.tsx, and lift the shared tone-style maps into
  data-screen.tone-styles.ts. Keeps the god-file ratchet green instead of
  growing an already-baselined monolith.
- Keep the destructive-action confirm dialog mounted while the action is in
  flight and ignore backdrop/Escape dismissal until it settles, so the
  loading state is visible and the operation can't be orphaned.
- Drop the redundant onKeyDown on the native "⋯" trigger; the button's
  intrinsic click handling already covers Enter/Space.

Added coverage for the in-flight confirm guard.
